+++ CHANGELOG.md
@@ -1,1 +1,8 @@
+## Build003 Release1
 
+### Added
+- FazerCards Provider foundation
+- Admin API Settings page
+- AJAX Test Connection
+- Account information display
+- Balance display